refactor: replace spinner with a custom one
