+++ src/components/BestBikes/BestBikes.tsx
@@ -1,11 +1,12 @@
 import { observer } from 'mobx-react-lite';
-import { Spin, Typography } from 'antd';
+import { Typography } from 'antd';
 import { useEffect } from 'react';
 import ProductCard from '../ProductCard/ProductCard';
 import { catalogStore } from '../../store/catalog-store';
 import styles from './BestBikes.module.css';
 import { BootState } from '../../types/boot-state';
 import { cartStore } from '../../store/cart-store';
+import { CubeSpin } from '../CubeSpinner/CubeSpinner';
 
 const { Title } = Typography;
 
@@ -21,7 +22,7 @@ export default observer(function BestBikes() {
       <Title level={2} className={styles['best-bikes-title']}>
         Best Selling Bikes
       </Title>
-      <Spin spinning={productsState === BootState.InProgress || cartStore.cartState === BootState.InProgress}>
+      <CubeSpin spinning={productsState === BootState.InProgress || cartStore.cartState === BootState.InProgress}>
         <ul className={styles['product-wrapper']}>
           {bestSellingProducts.map((product) => {
             return (
@@ -31,7 +32,7 @@ export default observer(function BestBikes() {
             );
           })}
         </ul>
-      </Spin>
+      </CubeSpin>
     </section>
   );
 });
