+++ src/components/LoginForm/LoginForm.tsx
@@ -1,4 +1,4 @@
-import { Form, Input, Typography, FormProps, notification, Spin } from 'antd';
+import { Form, Input, Typography, FormProps, notification } from 'antd';
 import type { NotificationArgsProps } from 'antd';
 import { FrownOutlined, SmileOutlined } from '@ant-design/icons';
 import { useNavigate } from 'react-router-dom';
@@ -8,6 +8,7 @@ import userStore from '../../store/user-store';
 import CustomButton from '../CustomButton/CustomButton';
 import { emailRules, passwordRules } from '../../utils/fields-validation';
 import { cartStore } from '../../store/cart-store';
+import { CubeSpin } from '../CubeSpinner/CubeSpinner';
 
 const { Text, Title, Link, Paragraph } = Typography;
 
@@ -75,7 +76,7 @@ export default function LoginForm() {
   };
 
   return (
-    <Spin spinning={spinRotate}>
+    <CubeSpin spinning={spinRotate}>
       <Form
         name="login"
         className={styles.loginForm}
@@ -124,6 +125,6 @@ export default function LoginForm() {
           </CustomButton>
         </Form.Item>
       </Form>
-    </Spin>
+    </CubeSpin>
   );
 }
