+++ src/components/ProductList/ProductList.tsx
@@ -1,4 +1,4 @@
-import { List, Spin } from 'antd';
+import { List } from 'antd';
 import { observer } from 'mobx-react-lite';
 import { useSearchParams } from 'react-router-dom';
 
@@ -8,6 +8,7 @@ import { cartStore } from '../../store/cart-store';
 import ProductCard from '../ProductCard/ProductCard';
 import styles from './ProductList.module.css';
 import { BootState } from '../../types/boot-state';
+import { CubeSpin } from '../CubeSpinner/CubeSpinner';
 
 export default observer(function ProductList() {
   const { products, productsState, totalPage, currentPage } = catalogStore;
@@ -23,7 +24,7 @@ export default observer(function ProductList() {
   }, []);
 
   return (
-    <Spin spinning={productsState === BootState.InProgress || cartStore.cartState === BootState.InProgress}>
+    <CubeSpin spinning={productsState !== BootState.Success || cartStore.cartState === BootState.InProgress}>
       <List
         bordered={false}
         itemLayout="vertical"
@@ -48,6 +49,6 @@ export default observer(function ProductList() {
         }}
         className={styles['product-list']}
       />
-    </Spin>
+    </CubeSpin>
   );
 });
