--- src/enums.ts
@@ -1,6 +1,0 @@
-export enum BootState {
-  None,
-  InProgress,
-  Success,
-  Failed,
-}
